+++ src/aliases.php
@@ -12,11 +12,19 @@ use Sentry\SentryBundle\EventListener\RequestListenerRequestEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerResponseEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerTerminateEvent;
 use Sentry\SentryBundle\EventListener\SubRequestListenerRequestEvent;
+use Sentry\SentryBundle\Tracing\Cache\TraceableCacheAdapter;
+use Sentry\SentryBundle\Tracing\Cache\TraceableCacheAdapterForV2;
+use Sentry\SentryBundle\Tracing\Cache\TraceableCacheAdapterForV3;
+use Sentry\SentryBundle\Tracing\Cache\TraceableTagAwareCacheAdapter;
+use Sentry\SentryBundle\Tracing\Cache\TraceableTagAwareCacheAdapterForV2;
+use Sentry\SentryBundle\Tracing\Cache\TraceableTagAwareCacheAdapterForV3;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\Compatibility\MiddlewareInterface;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingDriverForV2;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingDriverForV3;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingStatementForV2;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingStatementForV3;
+use Symfony\Component\Cache\Adapter\AdapterInterface;
+use Symfony\Component\Cache\DoctrineProvider;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
@@ -76,6 +84,26 @@ if (version_compare(Kernel::VERSION, '4.3.0', '>=')) {
 
     if (!class_exists(SubRequestListenerRequestEvent::class, false)) {
         class_alias(GetResponseEvent::class, SubRequestListenerRequestEvent::class);
+    }
+}
+
+if (interface_exists(AdapterInterface::class)) {
+    if (!class_exists(DoctrineProvider::class, false) && version_compare(\PHP_VERSION, '8.0.0', '>=')) {
+        if (!class_exists(TraceableCacheAdapter::class, false)) {
+            class_alias(TraceableCacheAdapterForV3::class, TraceableCacheAdapter::class);
+        }
+
+        if (!class_exists(TraceableTagAwareCacheAdapter::class, false)) {
+            class_alias(TraceableTagAwareCacheAdapterForV3::class, TraceableTagAwareCacheAdapter::class);
+        }
+    } else {
+        if (!class_exists(TraceableCacheAdapter::class, false)) {
+            class_alias(TraceableCacheAdapterForV2::class, TraceableCacheAdapter::class);
+        }
+
+        if (!class_exists(TraceableTagAwareCacheAdapter::class, false)) {
+            class_alias(TraceableTagAwareCacheAdapterForV2::class, TraceableTagAwareCacheAdapter::class);
+        }
     }
 }