+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add support for `symfony/cache-contracts` package version `3.x` (#588)
+
 ## 4.2.5 (2021-12-13)
 
 - Add support for Symfony 6 (#566)